fix: 예산 응답 스키마 Decimal→float 변환 (JSON 문자열 직렬화 버그 수정)
Pydantic Decimal 타입이 JSON에서 문자열로 직렬화되어
JS에서 '75830' > '750000' = true (문자열 비교)로 잘못 판단하던 버그.
BudgetOverviewResponse, CategoryBudgetResponse 응답 필드를 float으로 변경.

+++ backend/app/schemas/budget.py
@@ -14,20 +14,20 @@ class BudgetOverviewResponse(BaseModel):
     period_start: str
     period_end: str
     period_start_day: int
-    total_income: Decimal
-    total_fixed_expense: Decimal
-    total_transfer: Decimal
-    available_budget: Decimal
-    total_allocated: Decimal
-    unallocated: Decimal
+    total_income: float
+    total_fixed_expense: float
+    total_transfer: float
+    available_budget: float
+    total_allocated: float
+    unallocated: float
 
 
 class CategoryBudgetResponse(BaseModel):
     allocation_id: str
     category_id: str
-    allocated: Decimal
-    spent: Decimal
-    remaining: Decimal
+    allocated: float
+    spent: float
+    remaining: float
 
 
 class AllocationCreate(BaseModel):
